add support for images in dex-input
This commit also fixes the parsing of options by the FunctionCaller class.
When testing for the next char, the ConditionalEvaluator should
check both the output of the Lexer and the next char in the InputStream.
This is because the first char after a control sequence is processed by the
Lexer to complete the control sequence.